Ajout de tabs, debut tab carte
- Ajoute vue3-tabs-component, avec @custom_types directory pour
  declaration de types (manquantes pour cette dep)
- Split la legende a composantes reutilisables (pour le layout mobile):
  RegionSearch, Statistics. TODO: catastrophes et candidats
- Split du layout HTML (avec classes bootstrap pour faire un if/else)
  entre desktop et mobile, vu la difference significative entre les deux
  layouts maintenant
- Debut du layout de map, avec la map qui est un flex-grow pour prendre
  la place disponible sur le 100vh
- TODO: tabs catastrophes et candidat(e)s
- Placement du tooltip d'annee de recherche vers le bas, sinon ca cache
  le titre sur mobile
