+++ package.json
@@ -28,7 +28,8 @@
     "point-in-polygon": "^1.1.0",
     "vue": "^3.2.37",
     "vue-select": "^4.0.0-beta.3",
-    "vue-slider-component": "^4.1.0-beta.2"
+    "vue-slider-component": "^4.1.0-beta.2",
+    "vue3-tabs-component": "^1.1.0"
   },
   "devDependencies": {
     "@pinia/testing": "^0.0.13",
