+++ src/@custom_types/vue3-tabs-component.d.ts
@@ -1,0 +1,1 @@
+declare module 'vue3-tabs-component';
